README: re-generate, some manual fixes, add doc reference badge
The godoc2md tool has been deprecated, and is no longer maintained; it
has some bugs which prevent it from generating correct links (so I removed
those).

Added a badge to the documentation on pkg.go.dev, which perhaps should be
used as replacement for the docs in the README itself.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>